chore: Move deletion tasks to be co-located with the rest of deletions
I've moved files and imports only with this change. Celery task names
will be aligned with import paths separately.
