+++ src/sentry/deletions/defaults/__init__.py
@@ -2,6 +2,7 @@ from .alert_rule_trigger import *  # noqa: F401,F403
 from .alert_rule_trigger_action import *  # noqa: F401,F403
 from .alertrule import *  # noqa: F401,F403
 from .apiapplication import *  # noqa: F401,F403
+from .artifactbundle import *  # noqa: F401,F403
 from .commit import *  # noqa: F401,F403
 from .commitauthor import *  # noqa: F401,F403
 from .discoversavedquery import *  # noqa: F401,F403
