+++ src/sentry/deletions/tasks/groups.py
@@ -2,10 +2,10 @@ from collections.abc import Mapping, Sequence
 from typing import Any
 from uuid import uuid4
 
+from sentry.deletions.tasks.scheduled import MAX_RETRIES, logger
 from sentry.exceptions import DeleteAborted
 from sentry.silo.base import SiloMode
 from sentry.tasks.base import instrumented_task, retry, track_group_async_operation
-from sentry.tasks.deletion.scheduled import MAX_RETRIES, logger
 
 
 @instrumented_task(
