--- src/sentry/tasks/deletion/__init__.py
@@ -1,2 +1,0 @@
-from .groups import delete_groups  # noqa
-from .scheduled import reattempt_deletions, run_deletion, run_scheduled_deletions  # noqa
