+++ src/sentry/tasks/files.py
@@ -5,8 +5,9 @@ from django.utils import timezone
 
 from sentry.silo.base import SiloMode
 from sentry.tasks.base import instrumented_task, retry
-from sentry.tasks.deletion.scheduled import MAX_RETRIES
 from sentry.utils.db import atomic_transaction
+
+MAX_RETRIES = 5
 
 
 @instrumented_task(
