+++ tests/sentry/deletions/test_alert_rule.py
@@ -1,6 +1,6 @@
+from sentry.deletions.tasks.scheduled import run_scheduled_deletions
 from sentry.incidents.models.alert_rule import AlertRule, AlertRuleTrigger
 from sentry.models.organization import Organization
-from sentry.tasks.deletion.scheduled import run_scheduled_deletions
 from sentry.testutils.cases import TestCase
 from sentry.testutils.hybrid_cloud import HybridCloudTestMixin
