+++ tests/sentry/deletions/test_artifactbundle.py
@@ -1,3 +1,4 @@
+from sentry.deletions.tasks.scheduled import run_scheduled_deletions
 from sentry.models.artifactbundle import (
     ArtifactBundle,
     DebugIdArtifactBundle,
@@ -6,7 +7,6 @@ from sentry.models.artifactbundle import (
     SourceFileType,
 )
 from sentry.models.files.file import File
-from sentry.tasks.deletion.scheduled import run_scheduled_deletions
 from sentry.testutils.cases import TransactionTestCase
 from sentry.testutils.hybrid_cloud import HybridCloudTestMixin
