+++ tests/sentry/deletions/test_group.py
@@ -4,6 +4,7 @@ from uuid import uuid4
 
 from sentry import nodestore
 from sentry.deletions.defaults.group import EventDataDeletionTask
+from sentry.deletions.tasks.groups import delete_groups
 from sentry.eventstore.models import Event
 from sentry.models.eventattachment import EventAttachment
 from sentry.models.files.file import File
@@ -14,7 +15,6 @@ from sentry.models.grouphistory import GroupHistory, GroupHistoryStatus
 from sentry.models.groupmeta import GroupMeta
 from sentry.models.groupredirect import GroupRedirect
 from sentry.models.userreport import UserReport
-from sentry.tasks.deletion.groups import delete_groups
 from sentry.testutils.cases import SnubaTestCase, TestCase
 from sentry.testutils.helpers.datetime import before_now, iso_format
