+++ tests/sentry/deletions/test_sentry_app_installations.py
@@ -3,6 +3,7 @@ from django.db import router
 from django.db.transaction import get_connection
 
 from sentry import deletions
+from sentry.deletions.tasks.hybrid_cloud import schedule_hybrid_cloud_foreign_key_jobs
 from sentry.models.apigrant import ApiGrant
 from sentry.models.apitoken import ApiToken
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
@@ -13,7 +14,6 @@ from sentry.models.servicehook import ServiceHook
 from sentry.sentry_apps.installations import SentryAppInstallationCreator
 from sentry.silo.base import SiloMode
 from sentry.silo.safety import unguarded_write
-from sentry.tasks.deletion.hybrid_cloud import schedule_hybrid_cloud_foreign_key_jobs
 from sentry.testutils.cases import TestCase
 from sentry.testutils.outbox import outbox_runner
 from sentry.testutils.silo import assume_test_silo_mode, control_silo_test
