+++ tests/sentry/models/test_organization.py
@@ -13,6 +13,10 @@ from sentry.api.endpoints.organization_details import (
     update_tracked_data,
 )
 from sentry.auth.authenticators.totp import TotpInterface
+from sentry.deletions.tasks.hybrid_cloud import (
+    schedule_hybrid_cloud_foreign_key_jobs,
+    schedule_hybrid_cloud_foreign_key_jobs_control,
+)
 from sentry.models.apikey import ApiKey
 from sentry.models.auditlogentry import AuditLogEntry
 from sentry.models.notificationsettingoption import NotificationSettingOption
@@ -21,10 +25,6 @@ from sentry.models.options.organization_option import OrganizationOption
 from sentry.models.organization import Organization
 from sentry.models.organizationmember import OrganizationMember
 from sentry.silo.base import SiloMode
-from sentry.tasks.deletion.hybrid_cloud import (
-    schedule_hybrid_cloud_foreign_key_jobs,
-    schedule_hybrid_cloud_foreign_key_jobs_control,
-)
 from sentry.testutils.cases import TestCase
 from sentry.testutils.helpers.features import with_feature
 from sentry.testutils.hybrid_cloud import HybridCloudTestMixin
